Refactor user service and profile model imports for improved structure
